# python-flask-server-generated/gen/api/openapi.yaml
openapi: 3.0.0
info:
  description: API for managing beverages, labels, ingredients, container types, and
    unit measures in a beverage e-commerce platform.
  title: Beverage E-commerce API
  version: 1.0.0
servers:
- url: https://api.beverage-ecommerce.com/v2
security:
- BearerAuth:
  - read:beverages
  - write:beverages
  - read:brands
  - write:brands
  - read:labels
  - write:labels
  - read:ingredients
  - write:ingredients
  - read:container-types
  - write:container-types
...
paths:
...
  /brands:
    get:
      description: Retrieves a list of all brands.
      operationId: get_brands
      parameters:
      - description: Page number.
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - description: Number of items per page.
        explode: true
        in: query
        name: per_page
        required: false
        schema:
          default: 10
          maximum: 50
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Brand'
                type: array
                x-content-type: application/json
          description: A list of brands.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
      security:
      - BearerAuth:
        - read:brands
      summary: Retrieve all brand
      tags:
      - brand
      x-openapi-router-controller: swagger_server.controllers.brand_controller
      x-accepts: application/json
    post:
      description: Creates a new brand.
      operationId: create_brand
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/brands_body'
        required: true
      responses:
        "201":
          description: Brand created successfully.
          headers:
            Location:
              description: URL to the newly created brand.
              explode: false
              schema:
                format: uri
                type: string
              style: simple
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Access forbidden.
      security:
      - BearerAuth:
        - write:brands
        - read:brands
      summary: Create a new brand
      tags:
      - brand
      x-openapi-router-controller: swagger_server.controllers.brand_controller
      x-contentType: application/json
      x-accepts: application/json
  /brands/{id}:
    delete:
      description: Deletes a brand.
      operationId: delete_brand
      parameters:
      - description: Unique identifier for the brand.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Brand deleted successfully.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Access forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      security:
      - BearerAuth:
        - write:brands
        - read:brands
      summary: Delete a brand
      tags:
      - brand
      x-openapi-router-controller: swagger_server.controllers.brand_controller
      x-accepts: application/json
    get:
      description: Retrieves details of a specific brand.
      operationId: get_brand_by_id
      parameters:
      - description: Unique identifier for the brand.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Brand'
          description: Brand details.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      security:
      - BearerAuth:
        - read:brands
      summary: Retrieve a brand by ID
      tags:
      - brand
      x-openapi-router-controller: swagger_server.controllers.brand_controller
      x-accepts: application/json
    patch:
      description: Update an existing brand.
      operationId: update_brand
      parameters:
      - description: Unique identifier for the brand.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/brands_id_body_1'
        required: true
      responses:
        "200":
          description: Brand updated successfully.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Access forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      security:
      - BearerAuth:
        - write:brands
        - read:brands
      summary: Partially update a brand
      tags:
      - brand
      x-openapi-router-controller: swagger_server.controllers.brand_controller
      x-contentType: application/json
      x-accepts: application/json
    put:
      description: Create or replace an existing brand.
      operationId: replace_brand
      parameters:
      - description: Unique identifier for the brand.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/brands_id_body'
        required: true
      responses:
        "200":
          description: Brand updated successfully.
        "201":
          description: Brand created successfully.
          headers:
            Location:
              description: URL to the newly created brand.
              explode: false
              schema:
                format: uri
                type: string
              style: simple
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Access forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      security:
      - BearerAuth:
        - write:brands
        - read:brands
      summary: Update a brand
      tags:
      - brand
      x-openapi-router-controller: swagger_server.controllers.brand_controller
      x-contentType: application/json
      x-accepts: application/json
...
  /container-types:
    get:
      description: Retrieves a list of all container types.
      operationId: get_container_types
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ContainerType'
                type: array
                x-content-type: application/json
          description: A list of container types.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
      security:
      - BearerAuth:
        - read:container-types
      summary: Retrieve all container types
      tags:
      - container-type
      x-openapi-router-controller: swagger_server.controllers.container_type_controller
      x-accepts: application/json
    post:
      description: Creates a new container type.
      operationId: create_container_type
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/containertypes_body'
        required: true
      responses:
        "201":
          description: Container type created successfully.
          headers:
            Location:
              description: URL to the newly created container type.
              explode: false
              schema:
                format: uri
                type: string
              style: simple
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Access forbidden.
      security:
      - BearerAuth:
        - write:container-types
        - read:container-types
      summary: Create a new container type
      tags:
      - container-type
      x-openapi-router-controller: swagger_server.controllers.container_type_controller
      x-contentType: application/json
      x-accepts: application/json
  /container-types/{id}:
    delete:
      description: Deletes a container type.
      operationId: delete_container_type
      parameters:
      - description: Unique identifier for the container type.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Container type deleted successfully.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Access forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      security:
      - BearerAuth:
        - write:container-types
        - read:container-types
      summary: Delete a container type
      tags:
      - container-type
      x-openapi-router-controller: swagger_server.controllers.container_type_controller
      x-accepts: application/json
    get:
      description: Retrieves details of a specific container type.
      operationId: get_container_type_by_id
      parameters:
      - description: Unique identifier for the container type.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContainerType'
          description: Container type details.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      security:
      - BearerAuth:
        - read:container-types
      summary: Retrieve a container type by ID
      tags:
      - container-type
      x-openapi-router-controller: swagger_server.controllers.container_type_controller
      x-accepts: application/json
    patch:
      description: Updates an existing container type.
      operationId: update_container_type
      parameters:
      - description: Unique identifier for the container type.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/containertypes_id_body_1'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
          description: Container type updated successfully.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Access forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      security:
      - BearerAuth:
        - write:container-types
        - read:container-types
      summary: Update a container type
      tags:
      - container-type
      x-openapi-router-controller: swagger_server.controllers.container_type_controller
      x-contentType: application/json
      x-accepts: application/json
    put:
      description: Create or replace a container type.
      operationId: put_container_type
      parameters:
      - description: Unique identifier for the container type.
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/containertypes_id_body'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
          description: Container type updated successfully.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Access forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found.
      security:
      - BearerAuth:
        - write:container-types
        - read:container-types
      summary: Update a container type
      tags:
      - container-type
      x-openapi-router-controller: swagger_server.controllers.container_type_controller
      x-contentType: application/json
      x-accepts: application/json
components:
...
  schemas:
...
    Brand:
      example:
        brandName: brandName
        brandDescription: brandDescription
        brandId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        brandId:
          description: Unique identifier for the brand
          format: uuid
          type: string
        brandName:
          description: Name of the brand
          type: string
        brandDescription:
          description: Description of the brand
          type: string
      required:
      - brandId
      - brandName
      type: object
...
    ContainerType:
      example:
        volume: 0
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        type: type
      properties:
        id:
          description: Unique identifier for the container type.
          format: uuid
          type: string
        type:
          description: Type of container (e.g., Bottle, Can).
          type: string
        volume:
          description: Volume in milliliters.
          type: integer
      required:
      - id
      - type
      - volume
      type: object
...
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorResponse_error'
      type: object
...
    brands_body:
      properties:
        name:
          type: string
        description:
          type: string
      required:
      - name
      type: object
    brands_id_body:
      properties:
        name:
          type: string
        description:
          type: string
      required:
      - name
      type: object
    brands_id_body_1:
      properties:
        name:
          type: string
        description:
          type: string
      type: object
...
    containertypes_body:
      properties:
        type:
          type: string
        volume:
          format: float
          type: number
      required:
      - type
      - volume
      type: object
    containertypes_id_body:
      properties:
        type:
          type: string
        volume:
          format: float
          type: number
      required:
      - type
      - volume
      type: object
    inline_response_200_1:
      example:
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          type: string
      type: object
    containertypes_id_body_1:
      properties:
        type:
          type: string
        volume:
          format: float
          type: number
      type: object
    ErrorResponse_error:
      properties:
        code:
          description: HTTP status code.
          type: integer
        message:
          description: Error message.
          type: string
      required:
      - code
      - message
      type: object
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_BearerAuth